feat: add the ability to remove vertices, numerous features added
Add the ability to remove a vertex and all the edges associated with
it.

Update sokol headers to the latest version with a fix for sokol_gp,
waiting to merge upstream.

Render vertices as circles with a custom shader.

Big refactorization of the code, now it's a little bit more organized.